Added edit source helper
Necessary to combine functionality found in the constant source and method source helpers into a single helper. This adds an additional feature, the original helpers didn't have, which is opening source code in your default editor via the `EDITOR` environment variable.

The kit spec was updated to be dynamic when calculating the actual results with the expected results so we don't have to constantly update this spec as more helpers are added.

Milestone: minor