Added descender handler
Necessary to provide low level support for acquiring all descendants of a class based on current environment. `ObjectSpace` can be slow depending on how many objects are in memory. This also means the results can vary depending on what is loaded in memory and who is the subclass of the superclass you are investigating.

Milestone: minor