Updated to IRB 1.14.0
This version introduced the `cd` command which, sadly, breaks convention and doesn't use the `irb_` prefix which is now accounted for in the specs. Helper registration seems to have changed in this version as well. This is why the specs were updated to ensure the helpers are cleared each time when testing registration so there is no tainted data between the specs. link:https://github.com/ruby/irb/releases/tag/v1.14.0[Details].

Milestone: minor